requirements.txt
pillow
alive-progress